feat(core, refactor): refactor comparing baskets, find unique products and sum total price
