+++ basket_compare_playground/pl/jacek/services/apps/basketcompare/templates/product_search.html
@@ -46,7 +46,7 @@
         {% endif %}
     </div>
     <div class="col">
-        <a href="/basket_compare" class="btn btn-primary">Basket Compare</a>
+        <a href="/basket_compare" class="btn btn-primary" autofocus>Basket Compare</a>
     </div>
 </div>
