+++ tests/pl/jacek/services/apps/basketcompare/api/external/buybox/utils/test_buybox_data_utils.py
@@ -1,7 +1,7 @@
 from typing import List
 from unittest import TestCase
 
-from basket_compare_playground.pl.jacek.services.apps.basketcompare.api.external.buybox.utils.buybox_data_utils import \
+from basket_compare_playground.pl.jacek.services.apps.basketcompare.api.external.buybox.utils.model_utils import \
     sort_products_by_price
 from basket_compare_playground.pl.jacek.services.apps.basketcompare.api.external.buybox.model.Datum import Datum
